Add metrics syscalls back
