+++ crates/risc0-bonsai/src/host.rs
@@ -14,6 +14,7 @@ use risc0_zkvm::{
 };
 use sov_db::ledger_db::{LedgerDB, ProvingServiceLedgerOps};
 use sov_risc0_adapter::guest::Risc0Guest;
+use sov_risc0_adapter::host::add_benchmarking_callbacks;
 use sov_rollup_interface::zk::{Proof, Zkvm, ZkvmHost};
 use tracing::{error, info, warn};
 
@@ -314,7 +315,7 @@ impl<'a> ZkvmHost for Risc0BonsaiHost<'a> {
     /// Proofs are created on the Bonsai API.
     fn run(&mut self, with_proof: bool) -> Result<Proof, anyhow::Error> {
         if !with_proof {
-            let env = ExecutorEnvBuilder::default()
+            let env = add_benchmarking_callbacks(ExecutorEnvBuilder::default())
                 .write_slice(&self.env)
                 .build()
                 .unwrap();
