+++ crates/sovereign-sdk/adapters/risc0/src/lib.rs
@@ -8,6 +8,8 @@ use serde::{Deserialize, Serialize};
 use sov_rollup_interface::zk::Matches;
 
 pub mod guest;
+#[cfg(feature = "native")]
+pub mod host;
 
 #[cfg(feature = "bench")]
 pub mod metrics;
